impl: T-0005 — daemon logger + carried fixes + pattern promotion
   T-0005 CLOSED — verdict CONFIRMED (orchestrator, 2026-05-21).
   All 14 gate-blocking acceptance criteria pass; zero reactive deviations.

   - packages/daemon/src/log/logger.ts: first daemon source file
     - Promise-chain queue serializes writes (CC-1 satisfied)
     - Lazy file-handle open with mode 0o600 (CC-3)
     - Idempotent close() drains in-flight writes
     - stderr fallback on write failure
   - packages/daemon/src/index.ts: re-exports logger
   - packages/daemon/tests/logger.test.ts: 6 cases (4.a–4.f), all passing
   - packages/daemon/tsconfig.json: files: [] → include: ["src/**/*"]
   - packages/daemon/package.json: test/lint scripts + @types/node devDep
   - packages/shared/src/ipc.ts: daemon_uptime_s tightened to .int()
   - patterns/project/test-token-fixtures.md (created, active): codifies
     the inert-conforming-token rule observed in T-0003 + T-0004

   recommendedTypeChecked ran clean on first real async code — calibration
   signal validated. Total project tests: 30/30 across 4 files.

+++ packages/shared/src/ipc.ts
@@ -7,7 +7,7 @@ import { z } from "zod";
 export const StatusPayloadSchema = z
   .object({
     daemon_pid: z.number().int().nonnegative(),
-    daemon_uptime_s: z.number().nonnegative(),
+    daemon_uptime_s: z.number().int().nonnegative(),
     endpoint: z.string(),                       // e.g. "127.0.0.1:7423"
     tunnel_status: z.enum(["up", "degraded", "down"]),
     tunnel_url: z.string().nullable(),
